Separate from system libraries
That system folder is going to be getting quite large if I left all the sysmodule stuff in there.
