+++ src/core/libraries/libs.cpp
@@ -60,7 +60,7 @@
 #include "core/libraries/system/commondialog.h"
 #include "core/libraries/system/msgdialog.h"
 #include "core/libraries/system/posix.h"
-#include "core/libraries/system/sysmodule.h"
+#include "core/libraries/sysmodule/sysmodule.h"
 #include "core/libraries/system/systemservice.h"
 #include "core/libraries/system/userservice.h"
 #include "core/libraries/system_gesture/system_gesture.h"
