+++ src/core/libraries/sysmodule/sysmodule.cpp
@@ -1,17 +1,13 @@
-// SPDX-FileCopyrightText: Copyright 2025 shadPS4 Emulator Project
+// SPDX-FileCopyrightText: Copyright 2025-2026 shadPS4 Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
-
-#define MAGIC_ENUM_RANGE_MIN 0
-#define MAGIC_ENUM_RANGE_MAX 300
-#include <magic_enum/magic_enum.hpp>
 
 #include "common/logging/log.h"
 #include "core/libraries/error_codes.h"
 #include "core/libraries/kernel/process.h"
 #include "core/libraries/libs.h"
-#include "core/libraries/system/sysmodule.h"
-#include "core/libraries/system/system_error.h"
-#include "core/libraries/system/sysmodule_table.h"
+#include "core/libraries/sysmodule/sysmodule.h"
+#include "core/libraries/sysmodule/sysmodule_error.h"
+#include "core/libraries/sysmodule/sysmodule_table.h"
 
 namespace Libraries::SysModule {
 
@@ -68,7 +64,6 @@ int PS4_SYSV_ABI sceSysmoduleIsCameraPreloaded() {
 }
 
 int PS4_SYSV_ABI sceSysmoduleIsLoaded(OrbisSysModule id) {
-    // LOG_ERROR(Lib_SysModule, "(DUMMY) called module = {}", magic_enum::enum_name(id));
     if (static_cast<u16>(id) == 0) {
         LOG_ERROR(Lib_SysModule, "Invalid sysmodule ID: {:#x}", static_cast<u16>(id));
         return ORBIS_SYSMODULE_INVALID_ID;
@@ -77,7 +72,6 @@ int PS4_SYSV_ABI sceSysmoduleIsLoaded(OrbisSysModule id) {
 }
 
 int PS4_SYSV_ABI sceSysmoduleIsLoadedInternal(OrbisSysModuleInternal id) {
-    LOG_ERROR(Lib_SysModule, "(DUMMY) called module = {:#x}", static_cast<u32>(id));
     if ((static_cast<u32>(id) & 0x7FFFFFFF) == 0) {
         LOG_ERROR(Lib_SysModule, "Invalid internal sysmodule ID: {:#x}", static_cast<u32>(id));
         return ORBIS_SYSMODULE_INVALID_ID;
@@ -86,7 +80,6 @@ int PS4_SYSV_ABI sceSysmoduleIsLoadedInternal(OrbisSysModuleInternal id) {
 }
 
 int PS4_SYSV_ABI sceSysmoduleLoadModule(OrbisSysModule id) {
-    // LOG_ERROR(Lib_SysModule, "(DUMMY) called module = {}", magic_enum::enum_name(id));
     return ORBIS_OK;
 }
