+++ src/core/libraries/sysmodule/sysmodule_error.h
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: Copyright 2024-2026 shadPS4 Emulator Project
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#pragma once
+
+#include "common/types.h"
+
+constexpr s32 ORBIS_SYSMODULE_INVALID_ID = 0x805A1000;
+constexpr s32 ORBIS_SYSMODULE_NOT_LOADED = 0x805A1001;
+constexpr s32 ORBIS_SYSMODULE_LOCK_FAILED = 0x805A10FF;
