--- src/core/libraries/system/system_error.h
@@ -1,8 +1,0 @@
-// SPDX-FileCopyrightText: Copyright 2024 shadPS4 Emulator Project
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-#pragma once
-
-constexpr int ORBIS_SYSMODULE_INVALID_ID = 0x805A1000;
-constexpr int ORBIS_SYSMODULE_NOT_LOADED = 0x805A1001;
-constexpr int ORBIS_SYSMODULE_LOCK_FAILED = 0x805A10FF;
